GH-438: Move ResolutionException to *.utils as it is shared
