+++ protobuf-maven-plugin/src/main/java/io/github/ascopes/protobufmavenplugin/utils/SystemPathBinaryResolver.java
@@ -16,7 +16,6 @@
 
 package io.github.ascopes.protobufmavenplugin.utils;
 
-import io.github.ascopes.protobufmavenplugin.dependencies.ResolutionException;
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
